#2696 - Document-level recommendations
- Started moving extraction logic into the suggestion supports
