+++ inception/inception-recommendation/src/test/java/de/tudarmstadt/ukp/inception/recommendation/service/RecommendationServiceImplTest.java
@@ -17,16 +17,12 @@
  */
 package de.tudarmstadt.ukp.inception.recommendation.service;
 
-import static de.tudarmstadt.ukp.clarin.webanno.model.AnchoringMode.CHARACTERS;
-import static de.tudarmstadt.ukp.clarin.webanno.model.AnchoringMode.SENTENCES;
-import static de.tudarmstadt.ukp.clarin.webanno.model.AnchoringMode.SINGLE_TOKEN;
 import static de.tudarmstadt.ukp.clarin.webanno.model.AnchoringMode.TOKENS;
 import static de.tudarmstadt.ukp.inception.recommendation.api.RecommendationService.FEATURE_NAME_AUTO_ACCEPT_MODE_SUFFIX;
 import static de.tudarmstadt.ukp.inception.recommendation.api.RecommendationService.FEATURE_NAME_IS_PREDICTION;
 import static de.tudarmstadt.ukp.inception.recommendation.api.RecommendationService.FEATURE_NAME_SCORE_EXPLANATION_SUFFIX;
 import static de.tudarmstadt.ukp.inception.recommendation.api.RecommendationService.FEATURE_NAME_SCORE_SUFFIX;
 import static de.tudarmstadt.ukp.inception.recommendation.service.SuggestionExtraction.extractSuggestions;
-import static de.tudarmstadt.ukp.inception.recommendation.service.SuggestionExtraction.getOffsets;
 import static de.tudarmstadt.ukp.inception.support.uima.FeatureStructureBuilder.buildFS;
 import static java.util.Arrays.asList;
 import static org.apache.uima.cas.CAS.TYPE_NAME_ANNOTATION;
@@ -42,8 +38,6 @@ import static org.assertj.core.api.Assertions.tuple;
 import java.util.Arrays;
 
 import org.apache.uima.UIMAFramework;
-import org.apache.uima.fit.factory.CasFactory;
-import org.apache.uima.fit.factory.JCasFactory;
 import org.apache.uima.fit.testing.factory.TokenBuilder;
 import org.apache.uima.jcas.tcas.Annotation;
 import org.junit.jupiter.api.BeforeEach;
@@ -70,9 +64,6 @@ class RecommendationServiceImplTest
     private SourceDocument doc1;
     private SourceDocument doc2;
     private AnnotationLayer layer1;
-    private AnnotationLayer layer2;
-    private AnnotationFeature feature1;
-    private AnnotationFeature feature2;
 
     @BeforeEach
     void setup()
@@ -83,9 +74,6 @@ class RecommendationServiceImplTest
         doc2 = new SourceDocument("doc2", null, null);
         doc2.setId(2l);
         layer1 = AnnotationLayer.builder().withId(1l).withName("layer1").build();
-        layer2 = AnnotationLayer.builder().withId(2l).withName("layer2").build();
-        feature1 = AnnotationFeature.builder().withName("feat1").build();
-        feature2 = AnnotationFeature.builder().withName("feat2").build();
     }
 
     @Test
@@ -140,106 +128,6 @@ class RecommendationServiceImplTest
                 .extracting(AnnotationSuggestion::getId, AnnotationSuggestion::getLabel,
                         AnnotationSuggestion::getAge) //
                 .containsExactlyInAnyOrder(tuple(0, "aged", 1), tuple(3, "added", 0));
-    }
-
-    @Test
-    void testOffsetAlignmentWithAnchoringOnCharacters() throws Exception
-    {
-        var mode = CHARACTERS;
-        var targetCas = CasFactory.createCas();
-        tokenBuilder.buildTokens(targetCas.getJCas(), "This is a test .");
-
-        var suggestionCas = JCasFactory.createText(targetCas.getDocumentText());
-
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 0, 1))) //
-                .as("Trivial case: one character") //
-                .get().isEqualTo(new Offset(0, 1));
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 4, 8))) //
-                .as("Leading and trailing space should be removed") //
-                .get().isEqualTo(new Offset(5, 7));
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, -10, 100))) //
-                .as("Range should be clipped to document boundaries").get()
-                .isEqualTo(new Offset(0, targetCas.getDocumentText().length()));
-    }
-
-    @Test
-    void testOffsetAlignmentWithAnchoringOnSingleToken() throws Exception
-    {
-        var mode = SINGLE_TOKEN;
-        var targetCas = CasFactory.createCas();
-        tokenBuilder.buildTokens(targetCas.getJCas(), "This is a test .");
-
-        var suggestionCas = JCasFactory.createText(targetCas.getDocumentText());
-
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 0, 1))) //
-                .as("Reduce to empty if no token is fully covered") //
-                .isEmpty();
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 0, 4))) //
-                .as("Trivial case: one token") //
-                .get().isEqualTo(new Offset(0, 4));
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 0, 7))) //
-                .as("Discard suggestion if it covers more than one token") //
-                .isEmpty();
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, -10, 100))) //
-                .as("Discard suggestion if it covers more than one token (2)") //
-                .isEmpty();
-    }
-
-    @Test
-    void testOffsetAlignmentWithAnchoringOnTokens() throws Exception
-    {
-        var mode = TOKENS;
-        var targetCas = CasFactory.createCas();
-        tokenBuilder.buildTokens(targetCas.getJCas(), "This is a test .");
-
-        var suggestionCas = JCasFactory.createText(targetCas.getDocumentText());
-
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 0, 1))) //
-                .as("Reduce to empty if no token is fully covered") //
-                .isEmpty();
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 0, 4))) //
-                .as("Trivial case: one token") //
-                .get().isEqualTo(new Offset(0, 4));
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 0, 7))) //
-                .as("Trivial case: two tokens") //
-                .get().isEqualTo(new Offset(0, 7));
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 2, 12))) //
-                .as("Discard incompletely covered tokens") //
-                .get().isEqualTo(new Offset(5, 9));
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, -10, 100))) //
-                .as("Range should be clipped to document boundaries").get()
-                .isEqualTo(new Offset(0, targetCas.getDocumentText().length()));
-    }
-
-    @Test
-    void testOffsetAlignmentWithAnchoringOnSentences() throws Exception
-    {
-        var mode = SENTENCES;
-        var targetCas = CasFactory.createCas();
-        tokenBuilder.buildTokens(targetCas.getJCas(), "This is a test .\nAnother sentence here .");
-
-        var suggestionCas = JCasFactory.createText(targetCas.getDocumentText());
-        var sentences = targetCas.select(Sentence.class).asList();
-        var sent1 = sentences.get(0);
-        var sent2 = sentences.get(1);
-
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 0, 1))) //
-                .as("Reduce to empty if no sentence is fully covered") //
-                .isEmpty();
-        assertThat(getOffsets(mode, targetCas,
-                new Annotation(suggestionCas, sent1.getBegin(), sent1.getEnd()))) //
-                        .as("Trivial case: one sentence") //
-                        .get().isEqualTo(new Offset(sent1.getBegin(), sent1.getEnd()));
-        assertThat(getOffsets(mode, targetCas,
-                new Annotation(suggestionCas, sent1.getBegin(), sent2.getEnd()))) //
-                        .as("Trivial case: two sentences") //
-                        .get().isEqualTo(new Offset(sent1.getBegin(), sent2.getEnd()));
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, 0, 30))) //
-                .as("Discard incompletely covered sentences") //
-                .get().isEqualTo(new Offset(0, 16));
-        assertThat(getOffsets(mode, targetCas, new Annotation(suggestionCas, -10, 100))) //
-                .as("Range should be clipped to document boundaries").get()
-                .isEqualTo(new Offset(0, targetCas.getDocumentText().length()));
     }
 
     @Test
